vicell_kernel::sync

Struct Spinlock

pub struct Spinlock<T> {
    lock: AtomicBool,
    data: UnsafeCell<T>,
}

Simple spinlock.

Fields
lock: AtomicBooldata: UnsafeCell<T>
Implementations

impl<T> Spinlock<T>

pub const fn new(data: T) -> Self

pub fn lock(&self) -> SpinlockGuard<'_, T>

pub unsafe fn force_unlock(&self)

Force unlock the spinlock.

Safety

This is unsafe because it bypasses the lock guard. Should only be used in context switching or panic handlers.